requirements.txt
streamlit
openai
langchain
langchain_openai
langchainhub
streamlit-audiorecorder
google-auth
streamlit-oauth
pytest
isort
streamlit-authenticator
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
tzlocal
firebase_admin
streamlit_js_eval
googlemaps
setuptools
pyaudio
speechrecognition
pyttsx3
.